Add token generator to tokenGenerator.js
